changed CartRow component to stateless function, debugged 'cart view' by adding starting point of 0 to reduce function in cart.js
